refactor: 侧栏与内容区改用全局竖线分隔替代圆角卡片，提取 maskPhone 工具函数并统一导入路径
