+++ src/renderer/styles/variables.css
@@ -118,8 +118,8 @@
   --color-text-4: #c8c0bc;
 
   /* Border - 淡化一档（更贴近背景，柔和分隔） */
-  --border-default: #e6e2d8;
-  --border-light: #efece4;
+  --border-default: #dad4c6;
+  --border-light: #e6e2d8;
 
   /* Fill & Inverse */
   --fill: #f5f4ef;
