+++ src/renderer/utils/index.ts
@@ -14,3 +14,12 @@ export const removeStack = (...args: Array<() => void>) => {
 };
 
 export { uuid } from '@/common/utils';
+
+export function maskPhone(phone: string): string {
+  if (!phone) return '';
+  const digits = phone.replace(/\D/g, '');
+  if (digits.length === 11) {
+    return `${digits.slice(0, 3)}****${digits.slice(7)}`;
+  }
+  return phone;
+}
